Encode frame query parameters properly
The existing query parameter encoding was OK, but failed for a critical use-case:
graphite URLs containing colors that looked like #00AA00. The # character was causing
the URL to be truncated, as it has a special meaning in a URL.

The fix was not simple, because we display the unescaped URL to users (why?).
We are now manually parsing the query string instead of relying on shoving it into
an anchor tag. For our purposes, it should be fine; promdash doesn't do much w/ this.