Add URLParser library to handle serialization/deserialization of URLs
Since promdash has a few features for users that challenge normal url parsing
methods (shoving url in to an <a> tag, for example), this adds a little library
to handle finding the query and hash parts of the URL and properly encoding/decoding them.
In particular, there is support for treating the '#' character like part of a query string.
This is necessary for graphite urls, which may contain hexadecimal colors ;_;