User written functions for getting line color
Allow users to write functions that return the color of the line.  By
default, the function just returns the next color from the palette
(same as current behaviour).  Hoewver, the user can use the labels and
color palette object to find a new color.  For example:

    (function (labels, palette) {
      if (labels.bad === 'True') {
        return 'red';
      } else {
        return palette.color();
      }
    })

They can even use d3, which is one of the globals:

    (function (labels, palette) {
      if (labels.bad === 'True') {
        return d3.rgb(palette.color()).darker();
      } else {
        return palette.color();
      }
    })