[single widget] merge profile vars with dash vars
When a Profile was being used on a dashboard,
template variables were not being correctly merged
into it, erasing the template variables
completely.

This PR fixes the merge between the two different
variable types, instead of erasing the dashboard
variables.

The expected behavior of profile variables
overwriting dashboard variables is maintained.